+++ projects/js-packages/charts/src/charts/bar-chart/private/index.ts
@@ -1,1 +1,2 @@
 export { useBarChartOptions } from './use-bar-chart-options';
+export { TruncatedXTickComponent, TruncatedYTickComponent } from './truncated-tick-component';