+++ projects/js-packages/charts/tests/jest.config.cjs
@@ -4,11 +4,6 @@ const baseConfig = require( 'jetpack-js-tools/jest/config.base.js' );
 module.exports = {
 	...baseConfig,
 	rootDir: path.join( __dirname, '..' ),
-	setupFilesAfterEnv: [
-		...baseConfig.setupFilesAfterEnv,
-		'@testing-library/jest-dom',
-		'<rootDir>/tests/jest.setup.js',
-	],
 	transform: {
 		...baseConfig.transform,
 		'\\.[jt]sx?$': require( 'jetpack-js-tools/jest/babel-jest-config-factory.js' )(