+++ projects/packages/newsletter/src/class-settings.php
@@ -18,7 +18,7 @@ use Automattic\Jetpack\Status\Host;
  */
 class Settings {
 
-	const PACKAGE_VERSION = '0.1.4';
+	const PACKAGE_VERSION = '0.2.0';
 	/**
 	 * Whether the class has been initialized
 	 *