+++ projects/packages/publicize/_inc/components/customize-and-preview/tab-panels/mobile/index.tsx
@@ -1,8 +1,7 @@
 import { TabPanel } from '@wordpress/components';
 import clsx from 'clsx';
 import { useCallback } from 'react';
-import { CustomizationToggle } from '../../customization-toggle';
-import { usePerNetworkCustomization } from '../../customization-toggle/use-per-network-customization';
+import { usePerNetworkCustomization } from '../../../../hooks/use-per-network-customization';
 import { ConnectionTab } from '../types';
 import { useConnectionTabs } from '../use-connection-tabs';
 import { CustomizationSection } from './customization-section';
@@ -17,28 +16,28 @@ import { TabContent } from './tab-content';
 export function TabPanelMobile() {
 	const tabs = useConnectionTabs();
 
-	const { isEnabled: perNetwork } = usePerNetworkCustomization();
+	const { isEnabled: usingPerNetworkCustomization } = usePerNetworkCustomization();
 
 	const tabRenderer = useCallback(
 		( tab: ConnectionTab ) => {
-			return <TabContent connectionId={ tab.connectionId } perNetwork={ perNetwork } />;
+			return (
+				<TabContent
+					connectionId={ tab.connectionId }
+					usingPerNetworkCustomization={ usingPerNetworkCustomization }
+				/>
+			);
 		},
-		[ perNetwork ]
+		[ usingPerNetworkCustomization ]
 	);
 
 	return (
-		<div>
-			<div className={ styles[ 'customization-toggle-wrapper' ] }>
-				<CustomizationToggle />
-			</div>
-			<div className={ clsx( { [ styles[ 'tab-panel-mobile-wrapper-border' ] ]: ! perNetwork } ) }>
-				{ ! perNetwork && <CustomizationSection /> }
-				<TabPanel
-					className={ styles[ 'tab-panel-mobile' ] }
-					tabs={ tabs }
-					children={ tabRenderer }
-				/>
-			</div>
+		<div
+			className={ clsx( {
+				[ styles[ 'tab-panel-mobile-wrapper-border' ] ]: ! usingPerNetworkCustomization,
+			} ) }
+		>
+			{ ! usingPerNetworkCustomization && <CustomizationSection /> }
+			<TabPanel className={ styles[ 'tab-panel-mobile' ] } tabs={ tabs } children={ tabRenderer } />
 		</div>
 	);
 }