+++ projects/plugins/wpcomsh/wpcom-features/class-wpcom-features.php
@@ -499,6 +499,7 @@ class WPCOM_Features {
 	public const SUBSCRIBER_UNLIMITED_IMPORTS      = 'subscriber-unlimited-imports';
 	public const SUBSCRIPTION_GIFTING              = 'subscription-gifting';
 	public const SUPPORT                           = 'support';
+	public const TITAN_MAIL_1YEAR_TRIAL            = 'titan_mail_1year_trial';
 	public const UPGRADED_UPLOAD_FILETYPES         = 'upgraded_upload_filetypes';
 	public const UPLOAD_AUDIO_FILES                = 'upload-audio-files';
 	public const UPLOAD_PLUGINS                    = 'upload-plugins';
@@ -1045,6 +1046,10 @@ class WPCOM_Features {
 				'sticker_not_present' => 'gating-business-q1',
 				self::WPCOM_ALL_SITES,
 			),
+			array(
+				'sticker_not_present' => 'gating-business-q1',
+				self::WPCOM_PERSONAL_AND_HIGHER_PLANS,
+			),
 			self::WPCOM_PREMIUM_AND_HIGHER_PLANS,
 			self::JETPACK_ALL_SITES,
 		),
@@ -1052,6 +1057,10 @@ class WPCOM_Features {
 			array(
 				'sticker_not_present' => 'gating-business-q1',
 				self::WPCOM_ALL_SITES,
+			),
+			array(
+				'sticker_not_present' => 'gating-business-q1',
+				self::WPCOM_PERSONAL_AND_HIGHER_PLANS,
 			),
 			self::WPCOM_PREMIUM_AND_HIGHER_PLANS,
 			self::JETPACK_ALL_SITES,
@@ -1444,6 +1453,13 @@ class WPCOM_Features {
 			self::WPCOM_ALL_SITES,
 			self::JETPACK_PERSONAL_AND_HIGHER,
 			self::JETPACK_GOLDEN_TOKEN,
+		),
+		self::TITAN_MAIL_1YEAR_TRIAL            => array(
+			array(
+				'required_sticker' => 'gating-business-q1',
+				self::WPCOM_BUSINESS_PLANS,
+				self::WPCOM_ECOMMERCE_PLANS,
+			),
 		),
 		self::UPGRADED_UPLOAD_FILETYPES         => array(
 			self::SPACE_1GB,