+++ projects/js-packages/api/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [1.0.15] - 2026-01-19
+### Changed
+- Update package dependencies. [#46647]
+
 ## [1.0.14] - 2026-01-14
 ### Changed
 - Internal updates.
@@ -478,6 +482,7 @@
 - Add the API methods left behind by the previous PR.
 - Initial release of jetpack-api package
 
+[1.0.15]: https://github.com/Automattic/jetpack-api/compare/v1.0.14...v1.0.15
 [1.0.14]: https://github.com/Automattic/jetpack-api/compare/v1.0.13...v1.0.14
 [1.0.13]: https://github.com/Automattic/jetpack-api/compare/v1.0.12...v1.0.13
 [1.0.12]: https://github.com/Automattic/jetpack-api/compare/v1.0.11...v1.0.12