+++ projects/packages/sync/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.25.1] - 2026-01-19
+### Added
+- Add _wpas_connection_overrides and _wpas_customize_per_network to post meta whitelist for Jetpack Social. [#46563]
+
 ## [4.25.0] - 2026-01-14
 ### Added
 - Instant Search: Add 'guest' custom taxonomy for Sync. [#46562]
@@ -1624,6 +1628,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Move sync to a classmapped package
 
+[4.25.1]: https://github.com/Automattic/jetpack-sync/compare/v4.25.0...v4.25.1
 [4.25.0]: https://github.com/Automattic/jetpack-sync/compare/v4.24.2...v4.25.0
 [4.24.2]: https://github.com/Automattic/jetpack-sync/compare/v4.24.1...v4.24.2
 [4.24.1]: https://github.com/Automattic/jetpack-sync/compare/v4.24.0...v4.24.1